pyTest: organize tests in sub folders
